+++ packages/bpk-component-spinner/src/BpkExtraLargeSpinner.js
@@ -41,7 +41,7 @@ BpkExtraLargeSpinner.propTypes = {
 };
 
 BpkExtraLargeSpinner.defaultProps = {
-  type: SPINNER_TYPES.primary,
+  type: SPINNER_TYPES.dark,
   className: null,
 };