+++ packages/bpk-component-spinner/src/BpkLargeSpinner.js
@@ -49,7 +49,7 @@ BpkLargeSpinner.propTypes = {
 };
 
 BpkLargeSpinner.defaultProps = {
-  type: SPINNER_TYPES.primary,
+  type: SPINNER_TYPES.dark,
   className: null,
   alignToButton: false,
 };